chore(tests): harmonize Jest package version to address test file matching
This harmonizes the versions of Jest installed in Framework to address
an issue reported in Stencil here:
https://github.com/ionic-team/stencil/issues/3785

The issue is essentially that certain test commands, such as this one:

```
npx stencil test --spec --max-workers=2 src/utils/keyboard/test/
```

will, at present, alternate between running the tests as expected and
failing. Not good!

The issue is caused by having different versions of `@jest/core`, on the
one hand, and `jest` and `jest-cli` on the other. After updating `jest`
and `jest-cli` to match the version of `@jest/core` (`27.5.1`) and the
issue can no longer be reproduced.