Update interface.d.ts
I'd like to import `NavigationHookResult` from @ionic/core. In my case, to generate a typings file that allows SvelteKit developers to have a typings file consistent with the core-package. Now I am adding the `NavigationHookResult` type definition myself in the svelteHTML typings file.

This change allows me to import it and uses the same pattern for exporting as all other types