chore(item): modify slot deprecation notice to avoid JSDoc formatting issue (#26564)
This changes the deprecation notice for a few `@slot` tags on the
`ion-item` component. Previously the lines in question looked something
like this:

```ts
/**
 * @slot a slot that does something @deprecated this is going away!
 */
export class Item { ...
```

The intention behind formatting the JSDoc this way was to mark the
`@slot` itself as deprecated, to advice users that it will be removed in
the future.

Unfortunately, JSDoc tooling doesn't interpret the `@deprecated` tag as
referring to the `@slot` tag here, but instead as one of the tags
applied to the `Item` class. If we wanted to use `@deprecated` to mark
only a slot as deprecated then we'd need a way to write a JSDoc just for
that slot itself.

Since we don't want to mark the whole class as deprecated, as a
workaround, we can change to using `**DEPRECATED**` instead, which will
appear prominently in generated documentation but will not confuse tools
which are JSDoc-aware.