refactor(ci): allow manual 'stencil nightly' builds
this commit adds the `workflow_dispatch` event to trigger a stencil
nightly build workflow. this is beng done to make it easier to debug the
nightly stencil build, by allowing us to invoke it without opening a pr
against the framework repo.

this commit also renames the 'build-core' job to
'build-core-with-stencil-eval', to help differentiate the job associated
with this workflow from the 'build-core' job found in the primary ci
workflow