chore: upgrade project to angular 14
This commit bumps the developer dependencies and peer dependencies to Angular 14. As a result, we need to update ng-packagr, which requires the `ng-package.json` configuration instead of the config living local to the `package.json`. The platform file change is a build-time error with the types, as a result of the typescript dependency version bump.

There is a pending issue with the ng14 test app with this commit that I am digging into more and will fix in a follow-up commit.