chore(readme): remove experimental wording (#27398)
Issue number: N/A

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

We've been using the custom elements bundle in Ionic for a few years
now. The underlying behavior is stable in Stencil, so we should remove
the "experimental" verbiage from the README.


## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Removed experimental wording 
## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->