test(select, select-popover): migrate to generators (#27393)
Issue number: N/A

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

Select and select popover tests are using legacy syntax

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Select and select popover tests are using generator syntax


https://github.com/ionic-team/ionic-framework/pull/27393/commits/915bb49fadd4e1095cad9fef7d2abad71d3b07bc

- The select custom behavior does not vary across directions, so I
removed the extra RTL checks.


https://github.com/ionic-team/ionic-framework/pull/27393/commits/01555bd44b86d5d2a74a6a73205f819f2d60003b

- The single value screenshots were not being used, so I removed them.


https://github.com/ionic-team/ionic-framework/pull/27393/commits/8b4cffcaf899210d259f41e9dfeb9641e24afbe4

- The spec screenshots were not being used, so I removed them.


https://github.com/ionic-team/ionic-framework/pull/27393/commits/e50b08b2eddbf7d944c513bf8964c127eb3b0c18

- The legacy standalone check does not vary across directions, so I
removed the RTL checks.


https://github.com/ionic-team/ionic-framework/pull/27393/commits/43e26bab7719b5705c6418c675caf5d8633f6f1c

- The legacy single value tests do not vary across directions, so I
removed the the RTL checks


https://github.com/ionic-team/ionic-framework/pull/27393/commits/4935a68aab5a42837d9411d1bf305da244ef6b39

- The basic tests were creating screenshots of each overlay as they were
opened. However, we already have test coverage for how each overlay is
rendered in each overlay test suite. As a result, I removed the
screenshots in favor of a `toBeVisible` assertion. This ensures that
overlays are presented with the select, which is what I think we want to
prioritize testing.
- These same tests were also waiting for overlays to dismiss. This
behavior is already tested in the respective overlay test suites, so I
removed the dismiss calls.


https://github.com/ionic-team/ionic-framework/pull/27393/commits/8cc6c426f147c558b76aee169cc9f4fda97753b5
- The legacy basic tests were creating screenshots of each overlay as
they were opened. However, we already have test coverage for how each
overlay is rendered in each overlay test suite. As a result, I removed
the screenshots in favor of a `toBeVisible` assertion. This ensures that
overlays are presented with the select, which is what I think we want to
prioritize testing.
- These same tests were also waiting for overlays to dismiss. This
behavior is already tested in the respective overlay test suites, so I
removed the dismiss calls.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

---------

Co-authored-by: ionitron <hi@ionicframework.com>